refactor: cli utils under cli/
