+++ packages/fast-usdc/src/cli/util/agoric.js
