+++ packages/fast-usdc/src/cli/util/bank.js
