+++ packages/fast-usdc/src/cli/util/cctp.js
