+++ packages/fast-usdc/src/cli/util/file.js
