+++ packages/fast-usdc/src/cli/util/noble.js
