+++ packages/fast-usdc/test/util/file.test.ts
@@ -1,5 +1,5 @@
 import test from 'ava';
-import { makeFile } from '../../src/util/file.js';
+import { makeFile } from '../../src/cli/util/file.js';
 
 const makeReadMock = (content: string) => {
   let filePathRead: string;
